+++ kwcoco/coco_image.py
@@ -290,7 +290,7 @@ class CocoImage(ub.NiceRepr):
 
         # Choose "best" auxiliary image based on a hueristic.
         eye = kwimage.Affine.eye().matrix
-        asset_objs = img.get('auxiliary', img.get('assets', []))
+        asset_objs = img.get('auxiliary', img.get('assets', [])) or []
         for idx, obj in enumerate(asset_objs):
             # Take frobenius norm to get "distance" between transform and
             # the identity. We want to find the auxiliary closest to the
@@ -347,9 +347,9 @@ class CocoImage(ub.NiceRepr):
             # cant remove auxiliary otherwise inplace modification doesnt work
             # obj = ub.dict_diff(img, {'auxiliary'})
             yield obj
-        for obj in img.get('auxiliary', []):
+        for obj in img.get('auxiliary', None) or []:
             yield obj
-        for obj in img.get('assets', []):
+        for obj in img.get('assets', None) or []:
             yield obj
 
     def find_asset_obj(self, channels):
@@ -689,7 +689,7 @@ class CocoImage(ub.NiceRepr):
         img_info = _delay_load_imglike(bundle_dpath, img,
                                        nodata_method=nodata_method)
         obj_info_list = [(img_info, img)]
-        auxlist = img.get('auxiliary', img.get('assets', []))
+        auxlist = img.get('auxiliary', img.get('assets', [])) or []
         for aux in auxlist:
             info = _delay_load_imglike(bundle_dpath, aux,
                                        nodata_method=nodata_method)